requirements.txt
SQLAlchemy==2.0.30
python-dotenv==1.0.1
aiogram==3.7.0
validators==0.28.3
pendulum==3.0.0
aiocryptopay==0.4.1
cachetools==5.3.3
betterconf==2.6.1
ngrok==1.3.0
uvicorn==0.30.1
fastapi==0.111.0
APScheduler==3.10.4
pytz==2024.1
aiosqlite==0.20.0